refactor: move simple midi song + components to own module
Good template for the midi song builder
