+++ bevy_midix/src/song/builder/mod.rs
@@ -1,0 +1,7 @@
+/// A struct to build a [`MidiSong`] programatically.
+///
+/// This struct does not consider beats per minute or anything of that sort.
+///
+/// Straight micros.
+#[derive(Default, Clone, Debug)]
+pub struct MidiSongBuilder {}
