+++ bevy_midix/src/song/simple/channel_settings.rs
